Handle build case when gettextize is not available or not installed

+++ autogen.sh
@@ -12,10 +12,16 @@ fi
 
 set -x
 
+gettextize=$(which gettextize)
+if test -z "$gettextize"; then
+	echo "Couldn't find gettextize" >&2
+	exit 1
+fi
+
 # Copied from avahi's autogen.sh to work around gettext braindamage
 rm -f Makefile.am~ configure.ac~
 # Evil, evil, evil, evil hack
-sed 's/read dummy/\#/' `which gettextize` | sh -s -- --copy --force --no-changelog
+sed 's/read dummy/\#/' $gettextize | sh -s -- --copy --force --no-changelog
 test -f Makefile.am~ && mv Makefile.am~ Makefile.am
 test -f configure.ac~ && mv configure.ac~ configure.ac
